fix: building placement in parcels (#105)
DynamicCities determines the minimal and maximal parcel size by looking at the sizes of availbale buildings. The parcel size later gets encoded in the `shape` area.

When requiring the building size to be strictly smaller than the `shape` extents we end up with an "off-by-one" error.

This becomes most apparent if there's only a single building of fixed size, e.g., 10x10. During settlement generation we'll create parcels of size 10x10 (as that's the only available building we have anyway), but later check that the buildings to place are at max 9x9.

Doing an inclusive check with less-or-equals (≤) fixes this.